Renamed files "(.*).watch.list.(.*)" to "(.*).watchlist.(.*)"
