+++ src/doc_scan_service/index.js
@@ -8,7 +8,7 @@ const SdkConfigBuilder = require('./session/create/sdk.config.builder');
 const RequestedDocumentAuthenticityCheckBuilder = require('./session/create/check/requested.document.authenticity.check.builder');
 const RequestedIdDocumentComparisonCheckBuilder = require('./session/create/check/requested.id.document.comparison.check.builder');
 const RequestedThirdPartyIdentityCheckBuilder = require('./session/create/check/requested.third.party.identity.check.builder');
-const RequestedWatchlistScreeningCheckBuilder = require('./session/create/check/requested.watch.list.screening.check.builder');
+const RequestedWatchlistScreeningCheckBuilder = require('./session/create/check/requested.watchlist.screening.check.builder');
 const RequestedFaceMatchCheckBuilder = require('./session/create/check/requested.face.match.check.builder');
 const RequestedLivenessCheckBuilder = require('./session/create/check/requested.liveness.check.builder');
 const RequestedTextExtractionTaskBuilder = require('./session/create/task/requested.text.extraction.task.builder');
