+++ src/doc_scan_service/session/retrieve/watchlist.screening.check.response.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const WatchlistCheckResponse = require('./watch.list.check.response');
+const WatchlistCheckResponse = require('./watchlist.check.response');
 
 class WatchlistScreeningCheckResponse extends WatchlistCheckResponse {
 }
