+++ tests/doc_scan_service/session/retrieve/watchlist.summary.response.spec.js
@@ -1,4 +1,4 @@
-const WatchlistSummaryResponse = require('../../../../src/doc_scan_service/session/retrieve/watch.list.summary.response');
+const WatchlistSummaryResponse = require('../../../../src/doc_scan_service/session/retrieve/watchlist.summary.response');
 const RawResultsResponse = require('../../../../src/doc_scan_service/session/retrieve/raw.results.response');
 const SearchConfigResponse = require('../../../../src/doc_scan_service/session/retrieve/search.config.response');
